double instead float

+++ petuhov.kirill/T1/base-types.hpp
@@ -5,15 +5,15 @@ namespace petuhov
 {
   struct point_t
   {
-    float x;
-    float y;
+    double x;
+    double y;
   };
 
   struct rectangle_t
   {
     point_t pos;
-    float width;
-    float height;
+    double width;
+    double height;
   };
 }
