+++ petuhov.kirill/T1/rectangle.hpp
@@ -7,13 +7,13 @@ namespace petuhov
   class Rectangle : public Shape
   {
     public:
-      Rectangle(const petuhov::point_t &pos, float width, float height);
+      Rectangle(const petuhov::point_t &pos, double width, double height);
 
-      float getArea() const override;
+      double getArea() const override;
       petuhov::rectangle_t getFrameRect() const override;
       void move(const petuhov::point_t & pos) override;
-      void move(float dx, float dy) override;
-      void scale(const petuhov::point_t &center, float factor) override;
+      void move(double dx, double dy) override;
+      void scale(const petuhov::point_t &center, double factor) override;
 
     private:
       petuhov::rectangle_t rect_;
