+++ petuhov.kirill/T1/regular.cpp
@@ -2,7 +2,7 @@
 #include <stdexcept>
 #include <cmath>
 
-petuhov::Regular::Regular(const petuhov::point_t &pos, float radius, int vertexCount):
+petuhov::Regular::Regular(const petuhov::point_t &pos, double radius, int vertexCount):
   center_(pos),
   radius_(0.0f),
   vertexCount_(0)
@@ -25,15 +25,15 @@ petuhov::Regular::Regular(const petuhov::point_t &pos, float radius, int vertexC
   }
 }
 
-float petuhov::Regular::getArea() const
+double petuhov::Regular::getArea() const
 {
   return (vertexCount_ * radius_ * radius_ * sin(3.14 / vertexCount_)) / 2;
 }
 
 petuhov::rectangle_t petuhov::Regular::getFrameRect() const
 {
-  float width = 2 * radius_;
-  float height = 2 * radius_;
+  double width = 2 * radius_;
+  double height = 2 * radius_;
   return {center_, width, height};
 }
 
@@ -42,21 +42,21 @@ void petuhov::Regular::move(const petuhov::point_t &pos)
   center_ = pos;
 }
 
-void petuhov::Regular::move(float dx, float dy)
+void petuhov::Regular::move(double dx, double dy)
 {
   center_.x += dx;
   center_.y += dy;
 }
 
-void petuhov::Regular::scale(const petuhov::point_t &center, float factor)
+void petuhov::Regular::scale(const petuhov::point_t &center, double factor)
 {
   if (factor <= 0.0f)
   {
     throw std::invalid_argument("Invalid factor");
   }
 
-  float dx = center_.x - center.x;
-  float dy = center_.y - center.y;
+  double dx = center_.x - center.x;
+  double dy = center_.y - center.y;
 
   radius_ *= factor;
